avcodec/cfhd: Fix linesize type

+++ libavcodec/cfhd.c
@@ -808,7 +808,7 @@ static int cfhd_decode(AVCodecContext *avctx, void *data, int *got_frame,
         int lowpass_width   = s->plane[plane].band[0][0].width;
         int highpass_stride = s->plane[plane].band[0][1].stride;
         int act_plane = plane == 1 ? 2 : plane == 2 ? 1 : plane;
-        int dst_linesize;
+        ptrdiff_t dst_linesize;
         int16_t *low, *high, *output, *dst;
 
         if (avctx->pix_fmt == AV_PIX_FMT_BAYER_RGGB16) {
